Create a new module for ship drawing
